REFACTOR: reorganize src/infrastructure; get tests passing
