+++ dev-notes/12.0-StartupRecovery.md
@@ -16,23 +16,23 @@ On startup, I don't care about the result, but I want to add a route that allows
 ## What this process needs to do
 
 -  When startup reaches an appropriate point, publish a "ready for startup recovery" event and delays to let it start running
-   -  Delay will probably be about 1 second
--  The event subscriber runs the `RetryPendingRequestsUseCase`
+   -  Delay will probably be about 500ms by default
+-  The event subscriber runs the `RestartStalledRequestsUseCase`
 -  The use case
    -  gets all allowed request ids that were allowed before the event was created (pass event timestamp)
    -  gets all received request ids that were received before the event was created
-   -  add events for retry to a retry service
-   -  start retries
+   -  add events to a delayed event runner
+   -  start running events
    -  return results
--  The retry service
+-  The delayed event runner
    -  take the retry capability out of the circuit breaker and make it a standalone service
-   -  needs the events array, retry loop, ability to set delays, ability to stop retries, ability to start retries, ability to add events for retry
+   -  needs the events array, event runner loop, ability to set delays, ability to stop retries, ability to start retries, ability to add events for retry
 -  Effects on circuit breaker
-   -  add a retry service
-   -  use retry service's add event to add retries
-   -  use retry service's start retries when `awaitIsAvailable` gets available
+   -  add a delayed event runner
+   -  use delayed event runner's add event to add retries
+   -  use delayed event runner's start retries when `awaitIsAvailable` gets available
    -  when circuit breaker state changes, change retry delay based on state or stop retries (if it opens again)
--  The use case doesn't care if retries move from the use case's retry service to the circuit breaker's retry service
+-  The use case doesn't care if retries move from the use case's delayed event runner to the circuit breaker's delayed event runner
 
 ## Plan
