+++ dev-notes/12.4-SubscriberAndEvent.md
@@ -1,0 +1,31 @@
+# Event subscriber and event for startup
+
+## Planning
+
+I'll call the event `ApplicationStarted` so it can serve any other startup purpose. For example, other modules might have startup recovery in a separate use case and can subscribe to the same event. Publish the event after the application is built (all routes added). The publish happens in the server so controller tests aren't affected by it.
+
+-  If I included the publish in the application, controller tests would need to mock responses for the use case and know how to order them
+-  Ordering correctly is challenging given potential async behaviors, so keep it in the server
+-  Put the event in
+
+`ApplicationStartedSubscriber` will subscribe to the `ApplicationStarted` event and run the use case when called. Set up the use case and subscriber in the `initModule` method.
+
+-  Because the subscriber isn't tested as part of another test, write a test suite to ensure the subscriber runs
+-  Maybe add to the use case test as a separate `describe`
+-  Test is concerned about getting a response from the use case and handling ok and error conditions correctly
+
+The subscriber doesn't need to worry about connect errors. The delayed event runner is running events that go to other subscribers. Those subscribers are responsible to handle connect errors and add them to the appropriate repo's retries.
+
+## Reorganizing
+
+Thinking about where to put the event, I'm reorganizing `src/infrastructure`
+
+-  `resilience` directory will have the circuit breaker and delayed event runner code; event will go here
+-  `buildCircuitBreakers` will go into the `typeorm` or `prisma` directories because they're specific to those ORMs
+-  `logging` directory has general logging components (pino, winston)
+-  `AzureQueue` stays where it is forn ow
+-  I'm also leaving the Express and Fastify apps and servers in `src` for now
+
+All tests passing after correcting paths.
+
+**COMMIT: REFACTOR: reorganize src/infrastructure; get tests passing**
