+++ src/backup-request/adapter/impl/TypeormBackupRequestRepo.ts
@@ -1,7 +1,10 @@
 import { TypeormBackupRequest } from '../../../infrastructure/typeorm/entity/TypeormBackupRequest.entity';
 import { isTypeormConnectError } from '../../../infrastructure/typeorm/isTypeormConnectError';
 import { TypeormContext } from '../../../infrastructure/typeorm/typeormContext';
-import { CircuitBreakerWithRetry, ConnectFailureErrorData } from '../../../infrastructure/CircuitBreakerWithRetry';
+import {
+	CircuitBreakerWithRetry,
+	ConnectFailureErrorData,
+} from '../../../infrastructure/resilience/CircuitBreakerWithRetry';
 
 import { err, ok, Result } from '../../../common/core/Result';
 import { DomainEventBus } from '../../../common/domain/DomainEventBus';
