+++ src/backup-request/use-cases/check-request-allowed/CheckRequestAllowedUseCase.prisma.spec.ts
@@ -8,7 +8,7 @@ import { RequestTransportTypeValues } from '../../domain/RequestTransportType';
 import { CheckRequestAllowedDTO } from './CheckRequestAllowedDTO';
 import { CheckRequestAllowedUseCase } from './CheckRequestAllowedUseCase';
 
-import { CircuitBreakerWithRetry } from '../../../infrastructure/CircuitBreakerWithRetry';
+import { CircuitBreakerWithRetry } from '../../../infrastructure/resilience/CircuitBreakerWithRetry';
 import { ok } from '../../../common/core/Result';
 
 import {
