+++ src/backup-request/use-cases/check-request-allowed/CheckRequestAllowedUseCase.typeorm.spec.ts
@@ -1,4 +1,4 @@
-import { CircuitBreakerWithRetry } from '../../../infrastructure/CircuitBreakerWithRetry';
+import { CircuitBreakerWithRetry } from '../../../infrastructure/resilience/CircuitBreakerWithRetry';
 import { ok } from '../../../common/core/Result';
 import * as AdapterErrors from '../../../common/adapter/AdapterErrors';
