+++ src/backup-request/use-cases/restart-stalled-requests/RestartStalledRequestsUseCase.typeorm.spec.ts
@@ -1,6 +1,6 @@
 import { ok } from '../../../common/core/Result';
 
-import { CircuitBreakerWithRetry } from '../../../infrastructure/CircuitBreakerWithRetry';
+import { CircuitBreakerWithRetry } from '../../../infrastructure/resilience/CircuitBreakerWithRetry';
 import { TypeormBackupRequest } from '../../../infrastructure/typeorm/entity/TypeormBackupRequest.entity';
 import {
 	createMockTypeormContext,
