+++ src/expressServerTypeorm.ts
@@ -1,5 +1,5 @@
 import dotenv from 'dotenv';
-import { logger } from './infrastructure/pinoLogger';
+import { logger } from './infrastructure/logging/pinoLogger';
 
 const logContext = { location: 'Express+TypeORM', function: 'pre-start' };
 
@@ -14,7 +14,7 @@ dotenv.config({ path: `./env/${process.env.APP_ENV}.env` });
 
 import { typeormDataSource } from './infrastructure/typeorm/typeormDataSource';
 import { typeormCtx } from './infrastructure/typeorm/typeormContext';
-import { buildCircuitBreakers } from './infrastructure/buildCircuitBreakers.typeorm';
+import { buildCircuitBreakers } from './infrastructure/typeorm/buildCircuitBreakers.typeorm';
 
 import { buildApp } from './expressAppTypeorm';
