+++ src/fastifyApp.ts
@@ -3,7 +3,7 @@ import fastify from 'fastify';
 import { PrismaContext } from './infrastructure/prisma/prismaContext';
 
 import { addBackupRequestRoutes } from './backup-request/infrastructure/fastifyRoutes';
-import { ICircuitBreakers } from './infrastructure/buildCircuitBreakers.typeorm';
+import { ICircuitBreakers } from './infrastructure/prisma/buildCircuitBreakers.prisma';
 
 export function buildApp(prismaCtx: PrismaContext, circuitBreakers: ICircuitBreakers, opts: any = {}) {
 	const app = fastify(opts);
